+++ src/responses/Response.hpp
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   Response.hpp                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
+/*   By: emadriga <emadriga@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/01 20:44:32 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/05 20:50:38 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/05/06 09:33:25 by emadriga         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
  * @param status_code (int)
  * @param head_params (std::map <std::string, std::string>)
  * @param body (std::String)
- * 
+ *
 */
 class Response
 {
@@ -41,7 +41,5 @@ class Response
 		void return_error_message(int error_code);
 		std::string create_directory_index();
 };
-
-
 
 #endif